Migrate to static library builds for SDL3 and OpenAL Soft
Major dependency restructuring for improved build control and x64 support:

- Integrated SDL3 3.2.26 as static library with 234 source files
- Integrated OpenAL Soft 1.24.3 as static library with SDL3 backend
- Created SDL3.vcxproj for x64 Windows static library build
- Created openal-soft.vcxproj for x64 Windows static library build
- Updated HPL2.vcxproj to link against static libraries via project references
- Removed precompiled library dependencies (OpenAL32.lib, SDL3.lib)
- Added OpenAL headers: alext.h, efx-presets.h
- Added SDL3 build_config headers and GSL support library headers
- Updated .gitignore to exclude x64 build directories at all levels
- Updated README.md with project status

Technical improvements:
- All projects now build for x64 platform
- Full source control over critical dependencies
- Simplified dependency management
- Zero build errors and warnings
- Game runs successfully with new configuration

+++ HPL2/core/sources/impl/MouseSDL.cpp
@@ -96,7 +96,7 @@ namespace hpl {
 
 			if (pEvent->type == SDL_EVENT_MOUSE_MOTION)
 			{
-				mvMouseAbsPos = cVector2l(pEvent->motion.x, pEvent->motion.y);
+				mvMouseAbsPos = cVector2l(static_cast<int>(pEvent->motion.x), static_cast<int>(pEvent->motion.y));
 
 				// Button state tracking
 				// SDL3: state field still exists in motion event
